requirements.txt
asgiref==3.4.0; python_version >= "3.6"
django==3.2.4; python_version >= "3.6"
djangorestframework-simplejwt==4.7.1; python_version >= "3.7"
djangorestframework==3.12.4; python_version >= "3.5"
psycopg2-binary==2.9.1; python_version >= "3.6"
pyjwt==2.1.0; python_version >= "3.7"
pytz==2021.1; python_version >= "3.6"
sqlparse==0.4.1; python_version >= "3.6"
whitenoise==5.2.0; python_version >= "3.5" and python_version < "4"
